requirements.txt
pycryptodome>=3.8.2
requests>=2.19.1
six>=1.12.0
ordereddict>=1.1 ; python_version < '3.0'
